+++ app/src/test/java/com/ninjaapps/android/crosswords/puzzlemodel/CrosswordTemplateUnitTest.java
@@ -18,28 +18,28 @@ public class CrosswordTemplateUnitTest {
         byte height = 10;
 
         CrosswordTemplate template = new CrosswordTemplate(width, height);
-        template.setSeparator((byte)0, (byte)0);
-        template.setSeparator((byte)1, (byte)1);
-        template.setSeparator((byte)2, (byte)2);
-        template.setSeparator((byte)3, (byte)3);
-        template.setSeparator((byte)4, (byte)4);
-        template.setAt((byte)5, (byte)5, 'a');
-        template.setAt((byte)6, (byte)6, 'b');
-        template.setAt((byte)7, (byte)7, 'c');
-        template.setAt((byte)8, (byte)8, 'd');
-        template.setAt((byte)9, (byte)9, 'e');
+        template.setSeparator(0, 0);
+        template.setSeparator(1, 1);
+        template.setSeparator(2, 2);
+        template.setSeparator(3, 3);
+        template.setSeparator(4, 4);
+        template.setAt(5, 5, 'a');
+        template.setAt(6, 6, 'b');
+        template.setAt(7, 7, 'c');
+        template.setAt(8, 8, 'd');
+        template.setAt(9, 9, 'e');
 
-        assertTrue(template.isSeparator((byte)0, (byte)0));
-        assertTrue(template.isSeparator((byte)1, (byte)1));
-        assertTrue(template.isSeparator((byte)2, (byte)2));
-        assertTrue(template.isSeparator((byte)3, (byte)3));
-        assertTrue(template.isSeparator((byte)4, (byte)4));
+        assertTrue(template.isSeparator(0, 0));
+        assertTrue(template.isSeparator(1, 1));
+        assertTrue(template.isSeparator(2, 2));
+        assertTrue(template.isSeparator(3, 3));
+        assertTrue(template.isSeparator(4, 4));
 
-        assertEquals('a', template.getAt((byte)5, (byte)5));
-        assertEquals('b', template.getAt((byte)6, (byte)6));
-        assertEquals('c', template.getAt((byte)7, (byte)7));
-        assertEquals('d', template.getAt((byte)8, (byte)8));
-        assertEquals('e', template.getAt((byte)9, (byte)9));
+        assertEquals('a', template.getAt(5, 5));
+        assertEquals('b', template.getAt(6, 6));
+        assertEquals('c', template.getAt(7, 7));
+        assertEquals('d', template.getAt(8, 8));
+        assertEquals('e', template.getAt(9, 9));
 
         for (byte i = 0; i < width; ++i) {
             for (byte j = 0; j < height; ++j) {
@@ -58,17 +58,17 @@ public class CrosswordTemplateUnitTest {
         CrosswordTemplate template = new CrosswordTemplate(width, height);
 
         for (byte i = 0; i < height; i++) {
-            template.setSeparator((byte)(width + 1), i);
+            template.setSeparator((width + 1), i);
         }
         for (byte i = 0; i < width; i++) {
-            template.setAt(i, (byte)(height + 1), 'a');
+            template.setAt(i, (height + 1), 'a');
         }
 
         for (byte i = 0; i < height; i++) {
-            assertFalse(template.isSeparator((byte)(width + 1), i));
+            assertFalse(template.isSeparator((width + 1), i));
         }
         for (byte i = 0; i < width; i++) {
-            assertEquals(0, template.getAt(i, (byte)(height + 1)));
+            assertEquals(0, template.getAt(i, (height + 1)));
         }
     }
 
@@ -86,37 +86,37 @@ public class CrosswordTemplateUnitTest {
         */
 
         CrosswordTemplate template = new CrosswordTemplate(width, height);
-        template.setSeparator((byte)4, (byte)0);
-        template.setSeparator((byte)1, (byte)1);
-        template.setSeparator((byte)2, (byte)2);
-        template.setSeparator((byte)1, (byte)3);
+        template.setSeparator(4, 0);
+        template.setSeparator(1, 1);
+        template.setSeparator(2, 2);
+        template.setSeparator(1, 3);
 
-        assertEquals(4, template.getAvailableFrom((byte)0, (byte)0, true));
-        assertEquals(1, template.getAvailableFrom((byte)0, (byte)1, true));
-        assertEquals(2, template.getAvailableFrom((byte)0, (byte)2, true));
-        assertEquals(1, template.getAvailableFrom((byte)0, (byte)3, true));
-        assertEquals(5, template.getAvailableFrom((byte)0, (byte)4, true));
-        assertEquals(3, template.getAvailableFrom((byte)2, (byte)1, true));
-        assertEquals(2, template.getAvailableFrom((byte)3, (byte)2, true));
-        assertEquals(3, template.getAvailableFrom((byte)2, (byte)3, true));
+        assertEquals(4, template.getAvailableFrom(0, 0, true));
+        assertEquals(1, template.getAvailableFrom(0, 1, true));
+        assertEquals(2, template.getAvailableFrom(0, 2, true));
+        assertEquals(1, template.getAvailableFrom(0, 3, true));
+        assertEquals(5, template.getAvailableFrom(0, 4, true));
+        assertEquals(3, template.getAvailableFrom(2, 1, true));
+        assertEquals(2, template.getAvailableFrom(3, 2, true));
+        assertEquals(3, template.getAvailableFrom(2, 3, true));
 
-        assertEquals(5, template.getAvailableFrom((byte)0, (byte)0, false));
-        assertEquals(1, template.getAvailableFrom((byte)1, (byte)0, false));
-        assertEquals(2, template.getAvailableFrom((byte)2, (byte)0, false));
-        assertEquals(5, template.getAvailableFrom((byte)3, (byte)0, false));
-        assertEquals(1, template.getAvailableFrom((byte)1, (byte)2, false));
-        assertEquals(1, template.getAvailableFrom((byte)1, (byte)4, false));
-        assertEquals(2, template.getAvailableFrom((byte)2, (byte)3, false));
-        assertEquals(4, template.getAvailableFrom((byte)4, (byte)1, false));
+        assertEquals(5, template.getAvailableFrom(0, 0, false));
+        assertEquals(1, template.getAvailableFrom(1, 0, false));
+        assertEquals(2, template.getAvailableFrom(2, 0, false));
+        assertEquals(5, template.getAvailableFrom(3, 0, false));
+        assertEquals(1, template.getAvailableFrom(1, 2, false));
+        assertEquals(1, template.getAvailableFrom(1, 4, false));
+        assertEquals(2, template.getAvailableFrom(2, 3, false));
+        assertEquals(4, template.getAvailableFrom(4, 1, false));
 
-        assertEquals(0, template.getAvailableFrom((byte)4, (byte)0, true));
-        assertEquals(0, template.getAvailableFrom((byte)1, (byte)3, true));
-        assertEquals(0, template.getAvailableFrom((byte)1, (byte)1, true));
-        assertEquals(0, template.getAvailableFrom((byte)2, (byte)2, true));
-        assertEquals(0, template.getAvailableFrom((byte)4, (byte)0, false));
-        assertEquals(0, template.getAvailableFrom((byte)1, (byte)3, false));
-        assertEquals(0, template.getAvailableFrom((byte)1, (byte)1, false));
-        assertEquals(0, template.getAvailableFrom((byte)2, (byte)2, false));
+        assertEquals(0, template.getAvailableFrom(4, 0, true));
+        assertEquals(0, template.getAvailableFrom(1, 3, true));
+        assertEquals(0, template.getAvailableFrom(1, 1, true));
+        assertEquals(0, template.getAvailableFrom(2, 2, true));
+        assertEquals(0, template.getAvailableFrom(4, 0, false));
+        assertEquals(0, template.getAvailableFrom(1, 3, false));
+        assertEquals(0, template.getAvailableFrom(1, 1, false));
+        assertEquals(0, template.getAvailableFrom(2, 2, false));
     }
 
     @Test
@@ -133,13 +133,13 @@ public class CrosswordTemplateUnitTest {
         */
 
         CrosswordTemplate template = new CrosswordTemplate(width, height);
-        template.setSeparator((byte)4, (byte)0);
-        template.setSeparator((byte)1, (byte)1);
-        template.setSeparator((byte)2, (byte)2);
-        template.setSeparator((byte)1, (byte)3);
+        template.setSeparator(4, 0);
+        template.setSeparator(1, 1);
+        template.setSeparator(2, 2);
+        template.setSeparator(1, 3);
 
-        assertEquals(0, template.getAvailableFrom((byte)(width + 1), (byte)(height + 1), true));
-        assertEquals(0, template.getAvailableFrom((byte)(width + 1), (byte)(height + 1), false));
+        assertEquals(0, template.getAvailableFrom((width + 1), (height + 1), true));
+        assertEquals(0, template.getAvailableFrom((width + 1), (height + 1), false));
     }
 
     @Test
@@ -156,23 +156,23 @@ public class CrosswordTemplateUnitTest {
         */
 
         CrosswordTemplate template = new CrosswordTemplate(width, height);
-        template.setSeparator((byte)4, (byte)0);
-        template.setSeparator((byte)1, (byte)1);
-        template.setSeparator((byte)2, (byte)2);
-        template.setSeparator((byte)1, (byte)3);
+        template.setSeparator(4, 0);
+        template.setSeparator(1, 1);
+        template.setSeparator(2, 2);
+        template.setSeparator(1, 3);
 
-        assertEquals(0, template.getStartPoint((byte)3, (byte)0, true));
-        assertEquals(2, template.getStartPoint((byte)4, (byte)1, true));
-        assertEquals(3, template.getStartPoint((byte)4, (byte)2, true));
-        assertEquals(2, template.getStartPoint((byte)4, (byte)3, true));
-        assertEquals(0, template.getStartPoint((byte)4, (byte)4, true));
+        assertEquals(0, template.getStartPoint(3, 0, true));
+        assertEquals(2, template.getStartPoint(4, 1, true));
+        assertEquals(3, template.getStartPoint(4, 2, true));
+        assertEquals(2, template.getStartPoint(4, 3, true));
+        assertEquals(0, template.getStartPoint(4, 4, true));
 
-        assertEquals(0, template.getStartPoint((byte)0, (byte)4, false));
-        assertEquals(4, template.getStartPoint((byte)1, (byte)4, false));
-        assertEquals(2, template.getStartPoint((byte)1, (byte)2, false));
-        assertEquals(3, template.getStartPoint((byte)2, (byte)4, false));
-        assertEquals(0, template.getStartPoint((byte)3, (byte)4, false));
-        assertEquals(1, template.getStartPoint((byte)4, (byte)4, false));
+        assertEquals(0, template.getStartPoint(0, 4, false));
+        assertEquals(4, template.getStartPoint(1, 4, false));
+        assertEquals(2, template.getStartPoint(1, 2, false));
+        assertEquals(3, template.getStartPoint(2, 4, false));
+        assertEquals(0, template.getStartPoint(3, 4, false));
+        assertEquals(1, template.getStartPoint(4, 4, false));
     }
 
     @Test
@@ -189,13 +189,13 @@ public class CrosswordTemplateUnitTest {
         */
 
         CrosswordTemplate template = new CrosswordTemplate(width, height);
-        template.setSeparator((byte) 4, (byte) 0);
-        template.setSeparator((byte) 1, (byte) 1);
-        template.setSeparator((byte) 2, (byte) 2);
-        template.setSeparator((byte) 1, (byte) 3);
+        template.setSeparator( 4,  0);
+        template.setSeparator( 1,  1);
+        template.setSeparator( 2,  2);
+        template.setSeparator( 1,  3);
 
-        assertEquals(-1, template.getStartPoint((byte)1, (byte)1, true));
-        assertEquals(-1, template.getStartPoint((byte)(width + 1), (byte)(height + 1), true));
+        assertEquals(-1, template.getStartPoint(1, 1, true));
+        assertEquals(-1, template.getStartPoint((width + 1), (height + 1), true));
     }
 
     @Test
@@ -212,22 +212,22 @@ public class CrosswordTemplateUnitTest {
         */
 
         CrosswordTemplate template = new CrosswordTemplate(width, height);
-        template.setSeparator((byte) 4, (byte) 0);
-        template.setSeparator((byte) 1, (byte) 1);
-        template.setSeparator((byte) 2, (byte) 2);
-        template.setSeparator((byte) 1, (byte) 3);
+        template.setSeparator( 4,  0);
+        template.setSeparator( 1,  1);
+        template.setSeparator( 2,  2);
+        template.setSeparator( 1,  3);
 
-        assertEquals(1, template.getAvailableCount((byte)0, true));
-        assertEquals(1, template.getAvailableCount((byte)1, true));
-        assertEquals(2, template.getAvailableCount((byte)2, true));
-        assertEquals(1, template.getAvailableCount((byte)3, true));
-        assertEquals(1, template.getAvailableCount((byte)4, true));
+        assertEquals(1, template.getAvailableCount(0, true));
+        assertEquals(1, template.getAvailableCount(1, true));
+        assertEquals(2, template.getAvailableCount(2, true));
+        assertEquals(1, template.getAvailableCount(3, true));
+        assertEquals(1, template.getAvailableCount(4, true));
 
-        assertEquals(1, template.getAvailableCount((byte)0, false));
-        assertEquals(0, template.getAvailableCount((byte)1, false));
-        assertEquals(2, template.getAvailableCount((byte)2, false));
-        assertEquals(1, template.getAvailableCount((byte)3, false));
-        assertEquals(1, template.getAvailableCount((byte)4, false));
+        assertEquals(1, template.getAvailableCount(0, false));
+        assertEquals(0, template.getAvailableCount(1, false));
+        assertEquals(2, template.getAvailableCount(2, false));
+        assertEquals(1, template.getAvailableCount(3, false));
+        assertEquals(1, template.getAvailableCount(4, false));
     }
 
     @Test
@@ -244,13 +244,13 @@ public class CrosswordTemplateUnitTest {
         */
 
         CrosswordTemplate template = new CrosswordTemplate(width, height);
-        template.setSeparator((byte) 4, (byte) 0);
-        template.setSeparator((byte) 1, (byte) 1);
-        template.setSeparator((byte) 2, (byte) 2);
-        template.setSeparator((byte) 1, (byte) 3);
+        template.setSeparator( 4,  0);
+        template.setSeparator( 1,  1);
+        template.setSeparator( 2,  2);
+        template.setSeparator( 1,  3);
 
-        assertEquals(0, template.getAvailableCount((byte)(width + 1), true));
-        assertEquals(0, template.getAvailableCount((byte)(height + 1), false));
+        assertEquals(0, template.getAvailableCount((width + 1), true));
+        assertEquals(0, template.getAvailableCount((height + 1), false));
     }
 
     @Test
@@ -267,23 +267,23 @@ public class CrosswordTemplateUnitTest {
         */
 
         CrosswordTemplate template = new CrosswordTemplate(width, height);
-        template.setSeparator((byte) 4, (byte) 0);
-        template.setSeparator((byte) 1, (byte) 1);
-        template.setSeparator((byte) 2, (byte) 2);
-        template.setSeparator((byte) 1, (byte) 3);
+        template.setSeparator( 4,  0);
+        template.setSeparator( 1,  1);
+        template.setSeparator( 2,  2);
+        template.setSeparator( 1,  3);
 
-        assertTrue(template.setWord("fart", (byte)0, (byte)0, true));
-        assertTrue(template.setWord("arc", (byte)2, (byte)1, true));
-        assertTrue(template.setWord("re", (byte)0, (byte)2, true));
-        assertTrue(template.setWord("ua", (byte)3, (byte)2, true));
-        assertTrue(template.setWord("sen", (byte)2, (byte)3, true));
-        assertTrue(template.setWord("snore", (byte)0, (byte)4, true));
+        assertTrue(template.setWord("fart", 0, 0, true));
+        assertTrue(template.setWord("arc", 2, 1, true));
+        assertTrue(template.setWord("re", 0, 2, true));
+        assertTrue(template.setWord("ua", 3, 2, true));
+        assertTrue(template.setWord("sen", 2, 3, true));
+        assertTrue(template.setWord("snore", 0, 4, true));
 
-        assertTrue(template.setWord("forks", (byte)0, (byte)0, false));
-        assertTrue(template.setWord("ra", (byte)2, (byte)0, false));
-        assertTrue(template.setWord("so", (byte)2, (byte)3, false));
-        assertTrue(template.setWord("truer", (byte)3, (byte)0, false));
-        assertTrue(template.setWord("cane", (byte)4, (byte)1, false));
+        assertTrue(template.setWord("forks", 0, 0, false));
+        assertTrue(template.setWord("ra", 2, 0, false));
+        assertTrue(template.setWord("so", 2, 3, false));
+        assertTrue(template.setWord("truer", 3, 0, false));
+        assertTrue(template.setWord("cane", 4, 1, false));
     }
 
     @Test
@@ -300,20 +300,20 @@ public class CrosswordTemplateUnitTest {
         */
 
         CrosswordTemplate template = new CrosswordTemplate(width, height);
-        template.setSeparator((byte) 4, (byte) 0);
-        template.setSeparator((byte) 1, (byte) 1);
-        template.setSeparator((byte) 2, (byte) 2);
-        template.setSeparator((byte) 1, (byte) 3);
+        template.setSeparator( 4,  0);
+        template.setSeparator( 1,  1);
+        template.setSeparator( 2,  2);
+        template.setSeparator( 1,  3);
 
-        assertTrue(template.setWord("fart", (byte) 0, (byte) 0, true));
-        assertTrue(template.setWord("arc", (byte) 2, (byte) 1, true));
-        assertTrue(template.setWord("re", (byte) 0, (byte) 2, true));
-        assertTrue(template.setWord("ua", (byte) 3, (byte) 2, true));
-        assertTrue(template.setWord("sen", (byte) 2, (byte) 3, true));
-        assertTrue(template.setWord("snore", (byte) 0, (byte) 4, true));
+        assertTrue(template.setWord("fart",  0,  0, true));
+        assertTrue(template.setWord("arc",  2,  1, true));
+        assertTrue(template.setWord("re",  0,  2, true));
+        assertTrue(template.setWord("ua",  3,  2, true));
+        assertTrue(template.setWord("sen",  2,  3, true));
+        assertTrue(template.setWord("snore",  0,  4, true));
 
-        assertFalse(template.setWord("nonsense", (byte)0, (byte)0, false));
-        assertFalse(template.setWord("forks", (byte)0, (byte)1, false));
-        assertFalse(template.setWord("test", (byte)1, (byte)0, false));
+        assertFalse(template.setWord("nonsense", 0, 0, false));
+        assertFalse(template.setWord("forks", 0, 1, false));
+        assertFalse(template.setWord("test", 1, 0, false));
     }
 }
